refactor: Extract FileViewer And FilesViewerDrive
To make it easier to use on other views
